refactor: Explode files
