+++ src/ducks/pin/Round.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import styles from './styles.styl'
+
+const Round = props => {
+  return <button {...props} className={styles.Round} />
+}
+
+export default Round
