+++ src/ducks/pin/index.browser.jsx
@@ -1,0 +1,3 @@
+const pinGuarded = () => x => x
+
+export { pinGuarded }
