--- src/ducks/pin/index.jsx
@@ -1,2 +1,0 @@
-export { default as PinGuard } from './PinGuard'
-export { default as pinGuarded } from './hoc'
