+++ src/ducks/pin/index.mobile.jsx
@@ -1,0 +1,1 @@
+export { default as pinGuarded } from './hoc'
